+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-**0.14.2** (in progress; updated 13 Nov 2019) - *"Greenspun" [edition](https://en.wikipedia.org/wiki/Greenspun%27s_tenth_rule)*:
+**0.14.2** (in progress; updated 22 February 2020) - *"Greenspun" [edition](https://en.wikipedia.org/wiki/Greenspun%27s_tenth_rule)*:
 
 I think that with the arrival of [conditions and restarts](http://www.gigamonkeys.com/book/beyond-exception-handling-conditions-and-restarts.html), it is now fair to say `unpythonic` contains an ad-hoc, informally-specified, slow implementation of half of Common Lisp. To avoid *bug-ridden*, we have tests - but it's not entirely impossible for some to have slipped through.
 
@@ -30,6 +30,7 @@ If you still use 3.4 and find something in `unpythonic` doesn't work there, plea
   - `CountingIterator`: Count how many items have been yielded, as a side effect.
   - `slurp`: Extract all items from a `queue.Queue` (until it is empty) into a list, returning that list.
   - `map`: Curry-friendly thin wrapper for the builtin `map`, making it mandatory to specify at least one iterable.
+  - `running_minmax`, `minmax`: Extract both min and max in one pass over an iterable. The `running_` variant is a scan and returns a generator; the just-give-me-the-final-result variant is a fold.
 - `ulp`: Given a float `x`, return the value of the unit in the last place (the "least significant bit"). At `x = 1.0`, this is the [machine epsilon](https://en.wikipedia.org/wiki/Machine_epsilon), by definition of the machine epsilon.
 - `dyn` now supports rebinding, using the assignment syntax `dyn.x = 42`. For an atomic mass update, see `dyn.update`.
 - `box` now supports `.set(newvalue)` to rebind (returns the new value as a convenience), and `unbox(b)` to extract contents. Syntactic sugar for rebinding is `b << newvalue` (where `b` is a box).